added otp again for unverified user and add emp

+++ src/services/api-client.ts
@@ -15,7 +15,11 @@ const apiClient = axios.create({
 apiClient.interceptors.response.use(
   (response) => response,
   (error) => {
-    if (error.response && error.response.status === 401) {
+    if (
+      error.response &&
+      error.response.status === 401 &&
+      error.response.data.token
+    ) {
       localStorage.clear();
       window.location.href =
         (isProd
